rename vue pages in Maurice
